size_t

+++ include/aws/checksums/private/crc_util.h
@@ -10,12 +10,12 @@
 
 #define large_buffer_apply_impl(Name, T)                                                                               \
     T aws_large_buffer_apply_##Name(                                                                                   \
-        T (*checksum_fn)(const uint8_t *, int, T), const uint8_t *buffer, uint64_t length, T previous) {               \
+        T (*checksum_fn)(const uint8_t *, int, T), const uint8_t *buffer, size_t length, T previous) {               \
         T val = previous;                                                                                              \
         while (length > INT_MAX) {                                                                                     \
             val = checksum_fn(buffer, INT_MAX, val);                                                                   \
             buffer += (size_t)INT_MAX;                                                                                 \
-            length -= (uint64_t)INT_MAX;                                                                               \
+            length -= (size_t)INT_MAX;                                                                               \
         }                                                                                                              \
         val = checksum_fn(buffer, (int)length, val);                                                                   \
         return val;                                                                                                    \
